Move StoredToken to its own cpp/h
